Adding more error management

+++ scripts/docker.sh
@@ -17,12 +17,17 @@ rm -rf ${TMPDIR}
 
 docker system prune -f 2>&1 >/dev/null &
 
-if [ "x$1" = "xstart" ]; then
+if [ ! -d "${REPO_PATH}/node_modules" ]
+then
+  mkdir ${REPO_PATH}/node_modules 2>/dev/null || die "${REPO_PATH}/node_modules must exists and be a directory"
+fi
 
-  if [ ! -d "${REPO_PATH}/node_modules" ]
-  then
-    mkdir ${REPO_PATH}/node_modules 2>/dev/null || die "${REPO_PATH}/node_modules must exists and be a directory"
-  fi
+if [ ! -d "${REPO_PATH}/src/custom-elements/node_modules" ]
+then
+  mkdir ${REPO_PATH}/src/custom-elements/node_modules 2>/dev/null || die "${REPO_PATH}/src/custom-elements/node_modules must exists and be a directory"
+fi
+
+if [ "x$1" = "xstart" ]; then
 
   docker run --rm -it --name chi \
           --privileged \
